extract Config from OnlineIndexer
