+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexer.java
@@ -100,42 +100,13 @@ import static com.apple.foundationdb.record.provider.foundationdb.FDBRecordStore
 @API(API.Status.UNSTABLE)
 public class OnlineIndexer implements AutoCloseable {
     /**
-     * Default number of records to attempt to run in a single transaction.
-     */
-    public static final int DEFAULT_LIMIT = 100;
-    /**
-     * Default transaction write size limit. Note that the actual write might be "a little" bigger.
-     */
-    public static final int DEFAULT_WRITE_LIMIT_BYTES = 900_000;
-    /**
-     * Default limit to the number of records to attempt in a single second.
-     */
-    public static final int DEFAULT_RECORDS_PER_SECOND = 10_000;
-    /**
-     * Default number of times to retry a single range rebuild.
-     */
-    public static final int DEFAULT_MAX_RETRIES = 100;
-    /**
-     * Default interval to be logging successful progress in millis when building across transactions.
-     * {@code -1} means it will not log.
-     */
-    public static final int DEFAULT_PROGRESS_LOG_INTERVAL = -1;
-    /**
      * Default length between last access and lease's end time in milliseconds.
      */
     public static final long DEFAULT_LEASE_LENGTH_MILLIS = 10_000;
-    public static final long DEFAULT_TRANSACTION_TIME_LIMIT = 4_000;
     /**
      * Constant indicating that there should be no limit to some usually limited operation.
      */
     public static final int UNLIMITED = Integer.MAX_VALUE;
-
-    /**
-     * If {@link OnlineIndexer.Builder#getIncreaseLimitAfter()} is this value, the limit will not go back up, no matter how many
-     * successes there are.
-     * This is the default value.
-     */
-    public static final int DO_NOT_RE_INCREASE_LIMIT = -1;
 
     public static final int INDEXING_ATTEMPTS_RECURSION_LIMIT = 5; // Safety net - our algorithm should never reach this depth
 
@@ -154,7 +125,8 @@ public class OnlineIndexer implements AutoCloseable {
                   @Nonnull FDBRecordStore.Builder recordStoreBuilder,
                   @Nonnull List<Index> targetIndexes,
                   @Nullable Collection<RecordType> recordTypes,
-                  @Nullable UnaryOperator<Config> configLoader, @Nonnull Config config,
+                  @Nullable UnaryOperator<OnlineIndexOperationConfig> configLoader,
+                  @Nonnull OnlineIndexOperationConfig config,
                   boolean useSynchronizedSession,
                   long leaseLengthMillis,
                   boolean trackProgress,
@@ -396,7 +368,7 @@ public class OnlineIndexer implements AutoCloseable {
      */
     @Nonnull
     @VisibleForTesting
-    Config getConfig() {
+    OnlineIndexOperationConfig getConfig() {
         return common.config;
     }
 
@@ -861,374 +833,6 @@ public class OnlineIndexer implements AutoCloseable {
     }
 
     /**
-     * A holder for the mutable configuration parameters needed to rebuild an online index. These parameters are
-     * designed to be safe to be updated while a build is running.
-     */
-    @API(API.Status.UNSTABLE)
-    public static class Config {
-        private final int maxLimit;
-        private final int initialLimit;
-        private final int maxWriteLimitBytes;
-        private final int maxRetries;
-        private final int recordsPerSecond;
-        private final long progressLogIntervalMillis;
-        private final int increaseLimitAfter;
-        private final long timeLimitMilliseconds;
-        private final long transactionTimeLimitMilliseconds;
-        public static final long UNLIMITED_TIME = 0;
-
-        Config(int maxLimit, int initialLimit, int maxRetries, int recordsPerSecond, long progressLogIntervalMillis, int increaseLimitAfter,
-                   int maxWriteLimitBytes, long timeLimitMilliseconds, long transactionTimeLimitMilliseconds) {
-            this.maxLimit = maxLimit;
-            this.initialLimit = initialLimit;
-            this.maxRetries = maxRetries;
-            this.recordsPerSecond = recordsPerSecond;
-            this.progressLogIntervalMillis = progressLogIntervalMillis;
-            this.increaseLimitAfter = increaseLimitAfter;
-            this.maxWriteLimitBytes = maxWriteLimitBytes;
-            this.timeLimitMilliseconds = timeLimitMilliseconds;
-            this.transactionTimeLimitMilliseconds = transactionTimeLimitMilliseconds;
-        }
-
-        /**
-         * Get the maximum number of records to process in one transaction.
-         * @return the maximum number of records to process in one transaction
-         */
-        public int getMaxLimit() {
-            return maxLimit;
-        }
-
-        /**
-         * Get the initial number of records to process in one transaction.
-         * @return the initial number of records to process in one transaction
-         */
-        public int getInitialLimit() {
-            return initialLimit > 0 ? Math.min(initialLimit, maxLimit) : maxLimit;
-        }
-
-        /**
-         * Get the maximum number of times to retry a single range rebuild.
-         * @return the maximum number of times to retry a single range rebuild
-         */
-        public int getMaxRetries() {
-            return maxRetries;
-        }
-
-        /**
-         * Get the maximum number of records to process in a single second.
-         * @return the maximum number of records to process in a single second
-         */
-        public int getRecordsPerSecond() {
-            return recordsPerSecond;
-        }
-
-        /**
-         * Get the minimum time between successful progress logs when building across transactions.
-         * Negative will not log at all, 0 will log after every commit.
-         * @return the minimum time between successful progress logs in milliseconds
-         */
-        public long getProgressLogIntervalMillis() {
-            return progressLogIntervalMillis;
-        }
-
-        /**
-         * Get the number of successful range builds before re-increasing the number of records to process in a single
-         * transaction.
-         * By default this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
-         * @return the number of successful range builds before increasing the number of records processed in a single
-         * transaction
-         */
-        public int getIncreaseLimitAfter() {
-            return increaseLimitAfter;
-        }
-
-        /**
-         * Stop scanning if the write size (bytes) becomes bigger that this value.
-         * @return max write quota for a single transaction
-         */
-        public long getMaxWriteLimitBytes() {
-            return maxWriteLimitBytes;
-        }
-
-        /**
-         * Exit with exception if this limit is exceeded (checked after each non-final transaction).
-         * @return time limit in millisecond
-         */
-        public long getTimeLimitMilliseconds() {
-            return timeLimitMilliseconds;
-        }
-
-        /**
-         * Get the time quota for a single transaction.
-         * @return the time quota for a single transaction
-         */
-        public long getTransactionTimeLimitMilliseconds() {
-            return transactionTimeLimitMilliseconds;
-        }
-
-        @Nonnull
-        public static Builder newBuilder() {
-            return new Builder();
-        }
-
-        /**
-         * To create a builder for the given config.
-         * @return a {@link Config.Builder}
-         */
-        @Nonnull
-        public Builder toBuilder() {
-            return Config.newBuilder()
-                    .setMaxLimit(this.maxLimit)
-                    .setInitialLimit(this.initialLimit)
-                    .setWriteLimitBytes(this.maxWriteLimitBytes)
-                    .setIncreaseLimitAfter(this.increaseLimitAfter)
-                    .setProgressLogIntervalMillis(this.progressLogIntervalMillis)
-                    .setRecordsPerSecond(this.recordsPerSecond)
-                    .setMaxRetries(this.maxRetries)
-                    .setTimeLimitMilliseconds(timeLimitMilliseconds)
-                    .setTransactionTimeLimitMilliseconds(this.transactionTimeLimitMilliseconds);
-        }
-
-        /**
-         * A builder for {@link Config}. These are the mutable configuration parameters used while building indexes and are
-         * designed to be safe to be updated while a build is running.
-         */
-        @API(API.Status.UNSTABLE)
-        public static class Builder {
-            private int maxLimit = DEFAULT_LIMIT;
-            private int initialLimit = 0;
-            private int maxWriteLimitBytes = DEFAULT_WRITE_LIMIT_BYTES;
-            private int maxRetries = DEFAULT_MAX_RETRIES;
-            private int recordsPerSecond = DEFAULT_RECORDS_PER_SECOND;
-            private long progressLogIntervalMillis = DEFAULT_PROGRESS_LOG_INTERVAL;
-            private int increaseLimitAfter = DO_NOT_RE_INCREASE_LIMIT;
-            private long timeLimitMilliseconds = UNLIMITED_TIME;
-            private long transactionTimeLimitMilliseconds = DEFAULT_TRANSACTION_TIME_LIMIT;
-
-            protected Builder() {
-
-            }
-
-            /**
-             * Get the maximum number of records to process in one transaction.
-             * @return the maximum number of records to process in one transaction
-             * @see #setMaxLimit(int)
-             */
-            public int getMaxLimit() {
-                return maxLimit;
-            }
-
-            /**
-             * Set the maximum number of records to process in one transaction.
-             * The default limit is {@link #DEFAULT_LIMIT} = {@value #DEFAULT_LIMIT}.
-             * @param limit the maximum number of records to process in one transaction
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setMaxLimit(int limit) {
-                this.maxLimit = limit;
-                return this;
-            }
-
-            /**
-             * Get the initial number of records to process in one transaction.
-             * @return the initial number of records to process in one transaction
-             * @see #setInitialLimit(int)
-             */
-            public int getInitialLimit() {
-                return initialLimit;
-            }
-
-            /**
-             * Set the initial number of records to process in one transaction. This can be useful to avoid
-             * starting the indexing with the maximum limit (set by {@link #setMaxLimit(int)}), which may cause timeouts.
-             * The default initial limit is {@link #DEFAULT_LIMIT} = {@value #DEFAULT_LIMIT}.
-             * @param limit the initial number of records to process in one transaction
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setInitialLimit(int limit) {
-                this.initialLimit = limit;
-                return this;
-            }
-
-            /**
-             * Stop scanning if the write size (bytes) becomes bigger that this value.
-             * @return max write quota for a single transaction
-             * @see #setWriteLimitBytes(int)
-             */
-            public int getWriteLimitBytes() {
-                return maxWriteLimitBytes;
-            }
-
-            /**
-             * Set the maximum transaction size in a single transaction.
-             * The default limit is {@link #DEFAULT_WRITE_LIMIT_BYTES} = {@value #DEFAULT_WRITE_LIMIT_BYTES}.
-             * @param limit the approximate maximum write size in one transaction
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setWriteLimitBytes(int limit) {
-                this.maxWriteLimitBytes = limit;
-                return this;
-            }
-
-            /**
-             * Get the maximum number of times to retry a single range rebuild.
-             * @return the maximum number of times to retry a single range rebuild
-             * @see #setMaxRetries(int)
-             */
-            public int getMaxRetries() {
-                return maxRetries;
-            }
-
-            /**
-             * Set the maximum number of times to retry a single range rebuild.
-             *
-             * The default number of retries is {@link #DEFAULT_MAX_RETRIES} = {@value #DEFAULT_MAX_RETRIES}.
-             * @param maxRetries the maximum number of times to retry a single range rebuild
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setMaxRetries(int maxRetries) {
-                this.maxRetries = maxRetries;
-                return this;
-            }
-
-            /**
-             * Get the maximum number of records to process in a single second.
-             * @return the maximum number of records to process in a single second
-             * @see #setRecordsPerSecond(int)
-             */
-            public int getRecordsPerSecond() {
-                return recordsPerSecond;
-            }
-
-            /**
-             * Set the maximum number of records to process in a single second.
-             *
-             * The default number of retries is {@link #DEFAULT_RECORDS_PER_SECOND} = {@value #DEFAULT_RECORDS_PER_SECOND}.
-             * @param recordsPerSecond the maximum number of records to process in a single second
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setRecordsPerSecond(int recordsPerSecond) {
-                this.recordsPerSecond = recordsPerSecond;
-                return this;
-            }
-
-            /**
-             * Get the minimum time between successful progress logs when building across transactions.
-             * Negative will not log at all, 0 will log after every commit.
-             * @return the minimum time between successful progress logs in milliseconds
-             * @see #setProgressLogIntervalMillis(long)
-             */
-            public long getProgressLogIntervalMillis() {
-                return progressLogIntervalMillis;
-            }
-
-            /**
-             * Set the minimum time between successful progress logs when building across transactions.
-             * Negative will not log at all, 0 will log after every commit.
-             *
-             * @param progressLogIntervalMillis the number of milliseconds to wait between successful logs
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setProgressLogIntervalMillis(long progressLogIntervalMillis) {
-                this.progressLogIntervalMillis = progressLogIntervalMillis;
-                return this;
-            }
-
-            /**
-             * Get the number of successful range builds before re-increasing the number of records to process in a single
-             * transaction.
-             * By default, this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
-             * @return the number of successful range builds before increasing the number of records processed in a single
-             * transaction
-             * @see #setIncreaseLimitAfter(int)
-             */
-            public int getIncreaseLimitAfter() {
-                return increaseLimitAfter;
-            }
-
-            /**
-             * Set the number of successful range builds before re-increasing the number of records to process in a single
-             * transaction. The number of records to process in a single transaction will never go above {@link #getMaxLimit()}.
-             * By default this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
-             * @param increaseLimitAfter the number of successful range builds before increasing the number of records
-             * processed in a single transaction
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setIncreaseLimitAfter(int increaseLimitAfter) {
-                this.increaseLimitAfter = increaseLimitAfter;
-                return this;
-            }
-
-            /**
-             * Exit with exception if this limit is exceeded (checked after each non-final transaction).
-             * @return time limit in milliseconds
-             * @see #setTimeLimitMilliseconds(long)
-             */
-            public long getTimeLimitMilliseconds() {
-                return timeLimitMilliseconds;
-            }
-
-            /**
-             * Set the time limit. The indexer will exit with a proper exception if this time is exceeded after a
-             * non-final transaction.
-             * @param timeLimitMilliseconds the time limit in milliseconds
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setTimeLimitMilliseconds(long timeLimitMilliseconds) {
-                if (timeLimitMilliseconds < 0) {
-                    timeLimitMilliseconds = UNLIMITED_TIME;
-                }
-                this.timeLimitMilliseconds = timeLimitMilliseconds;
-                return this;
-            }
-
-            /**
-             * Get the time quota for a single transaction.
-             * @return the time quota for a single transaction
-             * @see #setTransactionTimeLimitMilliseconds(long)
-             */
-            public long getTransactionTimeLimitMilliseconds() {
-                return transactionTimeLimitMilliseconds;
-            }
-
-            /**
-             * Set the time limit for a single transaction. If this limit is exceeded, the indexer will commit the
-             * transaction and start a new one. This can be useful to avoid timeouts while scanning many records
-             * in each transaction.
-             * A non-positive value implies unlimited.
-             * Note that this limit, if reached, will be exceeded by an Order(1) overhead time. Keeping some margins might be
-             * a good idea.
-             * The default value is 4,000 (4 seconds), matches fdb's default 5 seconds transaction limit.
-             * @param timeLimitMilliseconds the time limit, per transaction, in milliseconds
-             * @return this builder
-             */
-            @Nonnull
-            public Builder setTransactionTimeLimitMilliseconds(long timeLimitMilliseconds) {
-                this.transactionTimeLimitMilliseconds = timeLimitMilliseconds;
-                return this;
-            }
-
-            /**
-             * Build a {@link Config}.
-             * @return a new Config object needed by {@link OnlineIndexer}
-             */
-            @Nonnull
-            public Config build() {
-                return new Config(maxLimit, initialLimit, maxRetries, recordsPerSecond, progressLogIntervalMillis, increaseLimitAfter,
-                        maxWriteLimitBytes, timeLimitMilliseconds, transactionTimeLimitMilliseconds);
-            }
-        }
-    }
-
-    /**
      * Builder for {@link OnlineIndexer}.
      *
      * <pre><code>
@@ -1458,7 +1062,7 @@ public class OnlineIndexer implements AutoCloseable {
         public OnlineIndexer build() {
             determineIndexingPolicy();
             validate();
-            Config conf = getConfig();
+            OnlineIndexOperationConfig conf = getConfig();
             return new OnlineIndexer(getRunner(), getRecordStoreBuilder(), targetIndexes, recordTypes,
                     getConfigLoader(), conf,
                     useSynchronizedSession, leaseLengthMillis, isTrackProgress(), indexingPolicy);
