+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexerSimpleTest.java
@@ -63,8 +63,8 @@ import java.util.stream.Collectors;
 import java.util.stream.LongStream;
 
 import static com.apple.foundationdb.record.metadata.Key.Expressions.field;
-import static com.apple.foundationdb.record.provider.foundationdb.OnlineIndexer.DEFAULT_PROGRESS_LOG_INTERVAL;
-import static com.apple.foundationdb.record.provider.foundationdb.OnlineIndexer.DO_NOT_RE_INCREASE_LIMIT;
+import static com.apple.foundationdb.record.provider.foundationdb.OnlineIndexOperationConfig.DEFAULT_PROGRESS_LOG_INTERVAL;
+import static com.apple.foundationdb.record.provider.foundationdb.OnlineIndexOperationConfig.DO_NOT_RE_INCREASE_LIMIT;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.instanceOf;
@@ -474,7 +474,7 @@ public class OnlineIndexerSimpleTest extends OnlineIndexerTest {
                 .setMdcContext(ImmutableMap.of("mdcKey", "my cool mdc value"))
                 .setMaxAttempts(3)
                 .setConfigLoader(old ->
-                        OnlineIndexer.Config.newBuilder()
+                        OnlineIndexOperationConfig.newBuilder()
                                 .setMaxLimit(100)
                                 .setMaxRetries(queue.size() + 3)
                                 .setRecordsPerSecond(10000)
@@ -999,7 +999,7 @@ public class OnlineIndexerSimpleTest extends OnlineIndexerTest {
 
     @Test
     void testIndexingThrottleBooker() {
-        final OnlineIndexer.Config config = OnlineIndexer.Config.newBuilder()
+        final OnlineIndexOperationConfig config = OnlineIndexOperationConfig.newBuilder()
                 .setInitialLimit(4)
                 .setRecordsPerSecond(100)
                 .setIncreaseLimitAfter(5)
@@ -1072,7 +1072,7 @@ public class OnlineIndexerSimpleTest extends OnlineIndexerTest {
 
     @Test
     void testIndexingThrottleBookerExceptions() {
-        final OnlineIndexer.Config config = OnlineIndexer.Config.newBuilder()
+        final OnlineIndexOperationConfig config = OnlineIndexOperationConfig.newBuilder()
                 .setInitialLimit(100)
                 .setRecordsPerSecond(100)
                 .setIncreaseLimitAfter(5)
@@ -1109,7 +1109,7 @@ public class OnlineIndexerSimpleTest extends OnlineIndexerTest {
 
     @Test
     void testIndexingThrottleBookerRepeatingExceptions() {
-        final OnlineIndexer.Config config = OnlineIndexer.Config.newBuilder()
+        final OnlineIndexOperationConfig config = OnlineIndexOperationConfig.newBuilder()
                 .setInitialLimit(1000)
                 .setRecordsPerSecond(100)
                 .setIncreaseLimitAfter(5)
